Move language variant method to L10nUtil
